// src/MapCore/Models/Enums/MapStillageType.ts
enum MapStillageType {
    NORMAL_TOP = 'NORMAL_TOP',
    NORMAL_BOTTOM = 'NORMAL_BOTTOM',
    NORMAL_LEFT_VERT = 'NORMAL_LEFT_VERT',
    NORMAL_RIGHT_VERT = 'NORMAL_RIGHT_VERT',

    SMALL_TOP = 'SMALL_TOP',
    SMALL_BOTTOM = 'SMALL_BOTTOM',
    SMALL_LEFT_VERT = 'SMALL_LEFT_VERT',
    SMALL_RIGHT_VERT = 'SMALL_RIGHT_VERT',

}

export default MapStillageType;